% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkVariable.r
\docType{methods}
\name{[}
\alias{[}
\alias{[,NetworkVariable-method}
\alias{[<-}
\alias{[<-,NetworkVariable-method}
\title{Extract or replace parts of a NetworkVariable object}
\description{
Extract or replace parts of a NetworkVariable object

set parts of NetworkVariable
}

