% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkMetadata.r
\name{idalter.gen}
\alias{idalter.gen}
\title{Generates IDs for people cited within an egocentric network}
\usage{
idalter.gen(metavar, nmetavarbyalter, metavar.id.alter = NULL)
}
\arguments{
\item{metavar}{\code{Rsocialdata} object, demographic variables for all individals of the egocentric network}

\item{nmetavarbyalter}{number of variable to use for each alter individual}

\item{metavar.id.alter}{if ids for alters are given en metavar we can use them, else we assume meta variables are given in the same order than id alters}
}
\description{
The function generates ID (\code{integer}) for flatten the network in a tabular view: instead of have one ID for ego and then an ID for each people of the network, we generate a unique ID for all people (ego and alters)
}

