% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkVariable.r
\name{is.squared.integer}
\alias{is.squared.integer}
\title{Test whether a number is the squared of an integer number}
\usage{
is.squared.integer(x)
}
\arguments{
\item{x}{an \code{integer}.}
}
\description{
The function returns \code{TRUE} if the root of the number supplied is an \code{integer}.
}

