% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allgenerics.r, R/measures.r
\docType{methods}
\name{net.centrality.degree}
\alias{net.centrality.degree}
\alias{net.centrality.degree,NetworkVariable-method}
\title{Compute network degree centrality measure}
\usage{
net.centrality.degree(x, target, which = "ego", direction = "in",
  quiet = FALSE)

\S4method{net.centrality.degree}{NetworkVariable}(x, target, which = "ego",
  direction = "in", quiet = FALSE)
}
\arguments{
\item{x}{the object for which we want to compute the centrality}

\item{target}{a \code{character} the label of the valid case to use for attesting of a link (see \code{valids(your_network)} or listing valid cases).}

\item{which}{a \code{character} if set to a node name, the centrality degree of this node is computed. If set to the value 'network', the network centrality degree is computed.}

\item{direction}{a \code{character} if set to 'in' the in-degree centrality is computed, if set to 'out' the out-degree centrality is computed.}

\item{quiet}{a \code{logical} if \code{TRUE} messages are disabled.}
}
\value{
A \code{ScaleVariable} object.
}
\description{
This generic method computes the centrality degree of a network object
}
\section{Methods (by class)}{
\itemize{
\item \code{NetworkVariable}: for \code{NetworkVariable} objects.
}}

