% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allgenerics.r, R/measures.r
\docType{methods}
\name{net.density}
\alias{net.density}
\alias{net.density,NetworkVariable-method}
\title{Compute network density measure}
\usage{
net.density(x, target, quiet = FALSE)

\S4method{net.density}{NetworkVariable}(x, target, quiet = FALSE)
}
\arguments{
\item{x}{the object for which we want to compute the density}

\item{target}{a \code{character} the label of the valid case to use for attesting of a link (see \code{valids(your_network)} or listing valid cases).}

\item{quiet}{a \code{logical} if \code{TRUE} messages are disabled.}
}
\value{
A \code{ScaleVariable} object.
}
\description{
This generic method computes the density of a network object
}
\section{Methods (by class)}{
\itemize{
\item \code{NetworkVariable}: method for \code{NetworkVariable} objects. The density is computed as the ratio of the number of existing connections over the total number of possible connections. The number of existing connections is computed as the number of occurences of the value '1' in the network. The total number of possible connections is computed as (number of row)*(number of rows - 1).
}}

