% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkMetadata.r
\name{netmetavar}
\alias{netmetavar}
\title{Creates NetworkMetavar objects}
\usage{
netmetavar(name, description, alterdata, names, nmetavarbyalter, nbalter,
  metavar.id.alter = NULL)
}
\arguments{
\item{name}{A \code{character} of length 1. Name of the \code{NetworkMetavar} object.}

\item{description}{A \code{character} of length 1. A short description for the \code{NetworkMetavar} object.}

\item{alterdata}{\code{Rsocialdata} object. The covariates variables.}

\item{names}{A \code{character}. Define names of the demographic variables describing people within networks.}

\item{nmetavarbyalter}{An \code{integer}. Number of variable to use for each alter individual.}

\item{nbalter}{An \code{integer}. Obsolete.}

\item{metavar.id.alter}{if An \code{integer}. IDs for alters are given en metavar we can use them, else we assume meta variables are given in the same order than id alters}
}
\description{
The function creates a code{NetworkMetavar} from a tabular view of the networks.
}

