% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkMetadata.r
\docType{class}
\name{NetworkMetadata-class}
\alias{NetworkMetadata-class}
\title{The NetworkMetadata class}
\description{
This object allows to store covariates data describing individuals within an egocentric networks. This covariates are generally demographic variables.
}

