% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allgenerics.r
\name{networks<-}
\alias{networks<-}
\title{Set the networks matrix of an object containing network data}
\usage{
networks(object) <- value
}
\arguments{
\item{object}{the \code{NetworkVariable} object for which we want to set the networks matrix.}

\item{value}{a list of network \code{matrix}}
}
\description{
This generic method intends to set the networks object.
}

