% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkVariable.r, R/allgenerics.r
\docType{methods}
\name{networks,NetworkVariable-method}
\alias{networks}
\alias{networks,NetworkVariable-method}
\title{Get/Set the networks matrix of an object containing network data}
\usage{
\S4method{networks}{NetworkVariable}(object)

networks(object)
}
\arguments{
\item{object}{the \code{NetworkVariable} object for which we want to get the networks matrix.}
}
\description{
This generic method intends to extract the networks object.
}
\section{Methods (by class)}{
\itemize{
\item \code{NetworkVariable}: method for \code{NetworkVariable} objects.
}}

