% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rsocialdata.network.r
\docType{package}
\name{Rsocialdata.network}
\alias{Rsocialdata.network}
\alias{Rsocialdata.network-package}
\title{Rsocialdata's network data toolbox}
\description{
Extension of the Rsocialdata package providing tools for handling network survey data
}
\keyword{Rsocialdata}
\keyword{networks}
\keyword{package}

