% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkVariable.r
\docType{class}
\name{NetworkVariable-class}
\alias{NetworkVariable-class}
\title{The NetworkVariable class}
\description{
This class allows to store network survey data. Basically, data are stored as matrix of survey codes and a dictionnary containing labels and missing values specification. The class supports a link to a NetworkMetavar object to take into account demographic information about people cited within the networks.
}

